Can now fetch both ingredients and recipes. Filter function works for a single ingredient so far. Filter with both test data and api data
